Refactor VM struct to increase register and call stack sizes
Increase the size of the `registers` array in the `VM` struct from 2000 to 8192 to accommodate more register values. Similarly, increase the size of the `call_stack` array from 16384 to 4096 to allow for a larger call stack.

This refactoring improves the performance and scalability of the VM by providing more space for storing register values and call frames.

Closes #123