Refactor: Add support for OP_FLOAT_MOD instruction
This commit adds support for the OP_FLOAT_MOD instruction in the VMThread struct. The OP_FLOAT_MOD instruction calculates the remainder of the division between two floating-point numbers and stores the result in a register. This change includes modifications to the create_dispatch_table function and the op_float_mod method in the VMThread struct.